+++ CHANGELOG.md
@@ -5,8 +5,9 @@ See [releases](https://github.com/heaths/go-vssetup/releases) for artifacts.
 
 ## v0.2.0
 
-* Removed `Locale()` that returned an LCID.
 * Added `UserPreferredLanguage()` to instead return a `language.Tag`.
+* Enabled color output if supported by the terminal.
+* Removed `Locale()` that returned an LCID.
 
 ## v0.1.1 (2021-09-11)